+++ plonk/simulation/__init__.py
@@ -4,7 +4,7 @@ It contains Plonk the implementation of smoothed particle hydrodynamics
 combined simulation data. This includes snapshots and time series data.
 """
 
-from .evolution import load_ev
 from .simulation import Simulation, load_sim
+from .time_series import load_time_series
 
-__all__ = ['Simulation', 'load_ev', 'load_sim']
+__all__ = ['Simulation', 'load_time_series', 'load_sim']